requirements-minimal.txt
pyyaml
service
